get_leagues_pipeline/requirements.txt
beautifulsoup4
requests
pandas
boto3
cloudscraper
pytest
pylint
psycopg2-binary